# services/gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Neobank Gateway API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /v1/notifications:
    get:
      operationId: listNotifications
      parameters:
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/XUserId"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
...
    Authorization:
      name: Authorization
      in: header
      schema:
        type: string
    XUserId:
      name: X-User-Id
      in: header
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Notification:
      type: object
      required: [id, user_id, event_type, title, body, read, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        event_type:
          type: string
        title:
          type: string
        body:
          type: string
        read:
          type: boolean
        created_at:
          type: string
          format: date-time
    NotificationList:
      type: object
      required: [notifications]
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
...